requirements.txt
blinker==1.9.0
boto3==1.39.6
botocore==1.39.6
click==8.2.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
